refactor(shared): split validation.ts into focused modules
Decompose the 728-LOC src/shared/utils/validation.ts into a slim public
facade plus focused sibling modules. Public API preserved via re-exports;
no behavior change.

- validationGuards.ts (119 LOC): raw shape interfaces + isValid* type guards.
- validationPlacement.ts (128 LOC): canPlaceBin (bounds/height/collision/zones).
- validationProperties.ts (86 LOC): validateCustomProperties.
- validationImport.ts (214 LOC): strict validateImport + PLACEMENT_REASON_MESSAGE.
- validationSalvage.ts (174 LOC): lenient salvageImport (moves bad bins to staging).
- validation.ts (109 LOC): facade re-exports + small helpers (clamp, truncate,
  formatDimension, validateLayoutIntegrity, getPlacementErrorMessage).
